moving back common classes to baker-interface module
