+++ core/baker-interface/pom.xml
@@ -26,11 +26,6 @@
             <artifactId>baker-annotations</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>com.ing.baker</groupId>
-            <artifactId>baker-runtime</artifactId>
-            <version>${project.version}</version>
-        </dependency>
 
         <dependency>
             <groupId>org.scala-lang</groupId>
