+++ core/baker-interface/src/main/scala/com/ing/baker/runtime/common/BakerException.scala
