+++ core/baker-interface/src/main/scala/com/ing/baker/runtime/common/InteractionExecutionFailureReason.java
