+++ core/baker-interface/src/main/scala/com/ing/baker/runtime/common/SensoryEventStatus.java
