+++ core/baker-interface/src/main/scala/com/ing/baker/runtime/model/BakerF.scala
@@ -2,14 +2,13 @@ package com.ing.baker.runtime.model
 
 import com.ing.baker.il.failurestrategy.ExceptionStrategyOutcome
 import com.ing.baker.il.{RecipeVisualStyle, RecipeVisualizer}
+import com.ing.baker.runtime.catseffect.{AsyncSupport, SyncSupport}
 import com.ing.baker.runtime.common
-import com.ing.baker.runtime.common.AsyncSupport.toAsync
-import com.ing.baker.runtime.common.AsyncSupport
+import com.ing.baker.runtime.catseffect.AsyncSupport.toAsync
 import com.ing.baker.runtime.common.FunctionK
 import com.ing.baker.runtime.common.BakerException.NoSuchIngredientException
 import com.ing.baker.runtime.common.LanguageDataStructures.ScalaApi
-import com.ing.baker.runtime.common.SyncSupport
-import com.ing.baker.runtime.common.SyncSupport.syntax._
+import SyncSupport.syntax._
 import com.ing.baker.runtime.common.{BakerException, InteractionExecutionFailureReason, RecipeRecord, SensoryEventStatus}
 import com.ing.baker.runtime.scaladsl.{Baker => DeprecatedBaker, _}
 import com.ing.baker.types.Value
