+++ pom.xml
@@ -107,7 +107,6 @@
         <!-- Core modules -->
         <module>core/baker-types</module>
         <module>core/baker-intermediate-language</module>
-        <module>core/baker-runtime</module>
         <module>core/baker-interface</module>
         <module>core/baker-interface-kotlin</module>
         <module>core/baker-interface-in-memory</module>
